There is no warnings on Aarch64 + Mac anymore.

+++ include/loops/loops.hpp
@@ -385,8 +385,8 @@ public:
 
     //TODO(ch): make next methods static:
     std::string getPlatformName() const;
-    size_t vbytes() const; //size of vector register in bytes
-    template<typename _Tp> inline size_t vlanes() const { return vbytes() / sizeof(_Tp); }
+    int vbytes() const; //size of vector register in bytes
+    template<typename _Tp> inline int vlanes() const { return vbytes() / (int)sizeof(_Tp); }
     void compileAll();
     void debugModeOn();//Gives ability to print listings even after compilation
 protected:
